+++ src/extensions/default/HTMLCodeHints/emmet-snippets.js
