+++ src/extensionsIntegrated/loader.js
@@ -43,5 +43,4 @@ define(function (require, exports, module) {
     require("./HtmlTagSyncEdit/main");
     require("./indentGuides/main");
     require("./CSSColorPreview/main");
-    require("./Emmet/main");
 });
